feat: Remove Terragrunt from development environment
This commit removes Terragrunt from the development environment configuration in the `flake.nix` file. Terragrunt is no longer needed, as the project is now using Terraform directly. The changes include:

- Removing Terragrunt from the list of allowed unfree packages in the `config.allowUnfreePredicate` section.
- Removing the `terragrunt version` command from the development shell script.
- Removing the pre-commit hook setup for Terragrunt, as it is no longer required.

The development environment now focuses on Terraform, OpenTofu, and other necessary development tools, simplifying the setup and maintenance of the project.