feat: Update Terraform version requirement to 1.10.0
This change updates the required Terraform version to 1.10.0 in the `main.tf` and `versions.tf` files. The previous version requirement of 1.5.0 and 1.3.0 has been updated to 1.10.0 to ensure compatibility with the latest Terraform version.

Additionally, the `aws` provider configuration has been removed from the `main.tf` file, as it is no longer needed for the module's functionality.

The `variables.tf` and `outputs.tf` files have also been updated to remove unnecessary variables and outputs, simplifying the module's configuration.