fix: Skip provider verification to avoid issues with provider caching
The changes in this commit address an issue with provider caching in the Terraform tests. The `TF_SKIP_PROVIDER_VERIFY` environment variable is set to `"1"` in the `SetupTerraformOptions`, `SetupModuleTerraformOptions`, and `SetupTargetTerraformOptions` functions. This skips the provider verification step, which can help avoid issues with provider caching and improve the reliability of the tests.

Additionally, the `.terraform.lock.hcl` files for the `basic` and `disabled_module` targets have been removed, as they are no longer needed.