fix: Ignore all .terraform.lock.hcl files within the modules/ directory
This commit adds a new ignore pattern to the .gitignore file to ignore all .terraform.lock.hcl files within the modules/ directory at any level. This is to ensure that these lock files are not accidentally committed to the repository, as they are specific to the local environment and should not be shared.